INFRA 1046 Move request processing folder
